refactor(core): extract overlay_scope_q to managers_overlay leaf (module-health cap)
managers.py exceeded the 500 LOC module-health cap after WP9 F1.6 added
overlay_scope_q. Move the overlay-scope Q-builders (overlay_scope_q +
_for_overlay) into a sibling managers_overlay.py leaf; re-import in
managers.py and point selectors._filters at the leaf directly.
