+++ src/teatree/core/managers.py
@@ -19,6 +19,8 @@ from teatree.core.loop_lease_manager import (
     is_per_loop_owner_slot,
     per_loop_owner_slot,
 )
+from teatree.core.managers_overlay import for_overlay as _for_overlay
+from teatree.core.managers_overlay import overlay_scope_q
 from teatree.core.repair_loop import IterationStalled, MaxIterationsExceeded
 from teatree.core.session_handover_manager import SessionHandoverManager, SessionHandoverQuerySet
 
@@ -76,40 +78,6 @@ def _claimable_now_q(now: datetime) -> Q:
     drift between "is there work" and the actual claim.
     """
     return Q(not_before__isnull=True) | Q(not_before__lte=now)
-
-
-def overlay_scope_q(overlay: str | None, *, prefix: str = "") -> Q:
-    """The Task-overlay scope clause as a reusable ``Q`` — the single source of truth.
-
-    A ``Task`` has no overlay column of its own: its overlay is its ticket's OR
-    its session's, so the clause spans both relations and always admits the
-    legacy empty-overlay rows (pre-multi-overlay data). ``prefix`` reaches the
-    ``ticket``/``session`` pair from a related model — ``"task__"`` scopes a
-    ``TaskAttempt`` by its task's overlay. An empty/``None`` overlay yields a
-    bare ``Q()`` that matches everything (``filter(Q())`` == ``all()``).
-
-    Shared by ``TaskQuerySet.for_overlay`` and the dashboard-selector filters
-    (``selectors._filters``) so the Task overlay clause can never drift between
-    the manager path and the read-model path (F1.6).
-    """
-    if not overlay:
-        return Q()
-    ticket = f"{prefix}ticket__overlay"
-    session = f"{prefix}session__overlay"
-    return Q(**{ticket: overlay}) | Q(**{session: overlay}) | Q(**{ticket: ""}) | Q(**{session: ""})
-
-
-def _for_overlay(qs: models.QuerySet, overlay: str | None) -> models.QuerySet:
-    """Scope *qs* to *overlay*, including legacy empty-overlay rows.
-
-    A module function rather than a mixin (composition over inheritance): the
-    three overlay-scoped QuerySets call it from their own ``for_overlay`` method,
-    so there is no mixin diamond and no ``# type: ignore[attr-defined]`` on
-    ``self.filter`` / ``self.all``.
-    """
-    if overlay:
-        return qs.filter(Q(overlay=overlay) | Q(overlay=""))
-    return qs.all()
 
 
 class TicketQuerySet(models.QuerySet):
